refactor: move General Settings into Settings folder
