+++ src/renderer/components/SettingsSection/General.tsx
